+++ fileHub/js-app/app/application.js
@@ -5,10 +5,10 @@ import {Router} from './services/router.js';
 import {RegistrationPage} from './register/registration-page.js';
 import {ErrorPage} from './ErrorPage.js';
 import {TitleService} from './services/title-service.js';
-import {FileListPage} from './user-page/file-list-page.js';
+import {FileListPage} from './file-list/file-list-page.js';
 import {StateManager} from './services/state-management/state-manager.js';
 import {ActionFactory} from './services/state-management/action-factory.js';
-import {RouteChanged} from './services/state-management/hash-changed-action/route-changed.js';
+import {RouteChanged} from './services/state-management/route-changed-action/route-changed.js';
 import {RoutingConfiguration} from './services/routing-configuration.js';
 import {mutator} from './services/state-management/mutator/mutator.js';
