+++ fileHub/js-app/app/file-list/file-list-item-view/file-list-item-view.js
