+++ fileHub/js-app/app/file-list/file-list-item-view/file-mime-types-list.js
