+++ fileHub/js-app/app/file-list/search-bar.js
