+++ fileHub/js-app/app/file-list/user-details.js
