+++ fileHub/js-app/app/services/state-management/action-factory.js
@@ -1,7 +1,7 @@
-import FetchCurrentFolder from './fetch-current-directory-action/fetch-current-folder.js';
-import FetchCurrentFolderExecutor from './fetch-current-directory-action/fetch-current-folder-executor.js';
-import {RouteChanged} from './hash-changed-action/route-changed.js';
-import {RouteChangedExecutor} from './hash-changed-action/route-changed-executor.js';
+import FetchCurrentFolder from './fetch-current-folder-action/fetch-current-folder.js';
+import FetchCurrentFolderExecutor from './fetch-current-folder-action/fetch-current-folder-executor.js';
+import {RouteChanged} from './route-changed-action/route-changed.js';
+import {RouteChangedExecutor} from './route-changed-action/route-changed-executor.js';
 import {GetRootFolder} from './get-root-folder-action/get-root-folder.js';
 import {GetRootFolderExecutor} from './get-root-folder-action/get-root-folder-executor.js';
 import {FetchCurrentFolderContent} from './fetch-current-folder-content-action/fetch-current-folder-content.js';
