+++ fileHub/js-app/app/services/state-management/fetch-current-folder-action/fetch-current-folder.js
